Change WithUlp for floats to use a short ulp error (not int)
It makes no sense to allow ULP comparisons for floats that are not very
close. A short sets a limit of approximately 2^15/2^24 = 2^-8 = 0.0039
of the full range of the mantissa for a relative error equivalent of
~3.8e-3.

Any comparisons above 32768 ULP must switch to using a relative
tolerance.

Note: For comparison the int ULP limit for double is 2^-21 of the 52 bit mantissa for relative error equivalent of 4.77e-7.