Add revapi allowed API changes
Changes are due to updated to the JUnit ExtensionContext class.
These are related to updated annotations and new classes returned from
additional methods that are part of interfaces that are input to the
API. These methods are not used and thus no breaking change will occur
when invoked with the previous JUnit version (5.6.2).